fixed winter category bug
The outdoor_cildren_widgets will duplicate content across categories in all outdoor themes. This fix while only applied to winter can be adapted to fix all other categories as well.

Currently this fix checks for the primary category that Yoast creates. If not it defaults to the first category in the list. And will only put it in th category it has set as primary.