adding plugins folder / GTM updates
this folder contains plugins that have custom code from the original plugins. These are regular plugins that need to go inside the defualt wordpress plugins folder, not the theme folder

The tag manager plugin needed updates to include a Page Type varibale, it was added to the api script and the datalayer script.